requirements.txt
streamlit
Pillow
imageio
numpy
